+++ requirements.txt
@@ -24,6 +24,7 @@ social-auth-app-django==4.0.0
 social-auth-core==3.3.3
 sqlparse==0.4.1
 urllib3==1.26.2
+django-image-optimizer==1.0.1
 django_seed
 django-tables2
 tablib
